Added upgrade-alps script

+++ base-system/extras/019-alps.sh
@@ -19,15 +19,27 @@ then
 
 cd /sources
 
-TARBALL="alps-new-$OS_VERSION.tar.gz"
+install_alps_from_tree() {
+    local tree="$1"
+    install -d /etc/alps /var/lib/alps /usr/bin
+    cp -a "$tree/usr/bin/alps" /usr/bin/alps
+    cp -a "$tree/etc/alps/." /etc/alps/
+    cp -a "$tree/var/lib/alps/." /var/lib/alps/
+    chmod a+x /usr/bin/alps
+    chmod -R a+rX /var/lib/alps/alps
+}
 
-mkdir -pv alps-extract
-tar xf $TARBALL -C alps-extract
-cp -r alps-extract/* /
-chmod a+x /usr/bin/alps
-chmod -R a+rX /var/lib/alps/alps
-
-rm -rf alps-extract
+if [ -f /sources/alps/usr/bin/alps ]; then
+    install_alps_from_tree /sources/alps
+else
+    TARBALL="alps-new-$OS_VERSION.tar.gz"
+    mkdir -pv alps-extract
+    tar xf "$TARBALL" -C alps-extract
+    cp -r alps-extract/* /
+    chmod a+x /usr/bin/alps
+    chmod -R a+rX /var/lib/alps/alps
+    rm -rf alps-extract
+fi
 
 mkdir -pv /var/cache/alps/{sources,packages,staging,ports}
 mkdir -pv /var/lib/alps/installed
